refactor: AsyncConfig에서 중복되는 코드를 BaseAsyncConfig로 통합
Resolves #34
