+++ admin/src/main/java/com/kjh/admin/config/AsyncConfig.java
@@ -1,31 +1,16 @@
 package com.kjh.admin.config;
 
-import com.kjh.core.exception.AsyncExceptionHandler;
-import org.springframework.aop.interceptor.AsyncUncaughtExceptionHandler;
+import com.kjh.core.config.BaseAsyncConfig;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.scheduling.annotation.AsyncConfigurer;
 import org.springframework.scheduling.annotation.EnableAsync;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
-
-import java.util.concurrent.Executor;
 
 @Configuration
 @EnableAsync
-public class AsyncConfig implements AsyncConfigurer {
+public class AsyncConfig {
 
-    @Override
-    public Executor getAsyncExecutor() {
-        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
-        executor.setCorePoolSize(5);
-        executor.setMaxPoolSize(10);
-        executor.setQueueCapacity(500);
-        executor.setThreadNamePrefix("Async-");
-        executor.initialize();
-        return executor;
-    }
-
-    @Override
-    public AsyncUncaughtExceptionHandler getAsyncUncaughtExceptionHandler() {
-        return new AsyncExceptionHandler();
+    @Bean
+    public BaseAsyncConfig baseAsyncConfig() {
+        return new BaseAsyncConfig(5, 10, 300, "Async-");
     }
 }
