+++ core/src/main/java/com/kjh/core/config/BaseAsyncConfig.java
@@ -1,0 +1,40 @@
+package com.kjh.core.config;
+
+import com.kjh.core.exception.AsyncExceptionHandler;
+import org.springframework.aop.interceptor.AsyncUncaughtExceptionHandler;
+import org.springframework.scheduling.annotation.AsyncConfigurer;
+import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
+
+import java.util.concurrent.Executor;
+
+public class BaseAsyncConfig implements AsyncConfigurer {
+
+    private final int corePoolSize;
+    private final int maxPoolSize;
+    private final int queueCapacity;
+    private final String threadNamePrefix;
+
+
+    public BaseAsyncConfig(int corePoolSize, int maxPoolSize, int queueCapacity, String threadNamePrefix) {
+        this.corePoolSize = corePoolSize;
+        this.maxPoolSize = maxPoolSize;
+        this.queueCapacity = queueCapacity;
+        this.threadNamePrefix = threadNamePrefix;
+    }
+
+    @Override
+    public Executor getAsyncExecutor() {
+        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
+        executor.setCorePoolSize(corePoolSize);
+        executor.setMaxPoolSize(maxPoolSize);
+        executor.setQueueCapacity(queueCapacity);
+        executor.setThreadNamePrefix(threadNamePrefix);
+        executor.initialize();
+        return executor;
+    }
+
+    @Override
+    public AsyncUncaughtExceptionHandler getAsyncUncaughtExceptionHandler() {
+        return new AsyncExceptionHandler();
+    }
+}
